Implement cursor links for records
When multiple records are returned there are links embedded in them.
These links provide pagination cursors to the next and previous page.
We now parse these values out of the uris and provide them in order to
page to the next end point.
