+++ client/src/endpoint/mod.rs
@@ -1,12 +1,13 @@
 //! This module contains the various end point definitions for stellar's horizon
 //! API server.
 use error::Result;
-use std;
-use serde::de::{Deserialize, DeserializeOwned, Deserializer};
+use serde::de::DeserializeOwned;
 use http;
 
 mod account;
 mod asset;
+mod records;
+pub use self::records::Records;
 pub use self::account::AccountDetails;
 pub use self::asset::AllAssets;
 
@@ -36,80 +37,5 @@ impl Order {
             Order::Asc => "asc".to_string(),
             Order::Desc => "desc".to_string(),
         }
-    }
-}
-
-/// A struct that represents a set of records returned from the horizon api.
-#[derive(Debug)]
-pub struct Records<T>
-where
-    T: DeserializeOwned,
-{
-    records: Vec<T>,
-}
-
-impl<T> Records<T>
-where
-    T: DeserializeOwned,
-{
-    /// Returns a slice of the embedded records.
-    pub fn records<'a>(&'a self) -> &'a Vec<T> {
-        &self.records
-    }
-}
-
-impl<'de, T> Deserialize<'de> for Records<T>
-where
-    T: DeserializeOwned,
-{
-    fn deserialize<D>(d: D) -> std::result::Result<Records<T>, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let embedded: Embedded<RecordsIntermediate<T>> = Embedded::deserialize(d)?;
-        Ok(Records {
-            records: embedded.embedded.records,
-        })
-    }
-}
-
-/// The HAL response format will embed resources within it. When it does
-/// this provides a wrapper to the `_embedded` key.
-///
-/// https://www.stellar.org/developers/horizon/reference/responses.html
-#[derive(Deserialize)]
-struct Embedded<T> {
-    #[serde(rename = "_embedded")] embedded: T,
-}
-
-/// If the embedded resource is a set of records, this can provide that data back in
-/// a generic way.
-#[derive(Deserialize)]
-struct RecordsIntermediate<T> {
-    records: Vec<T>,
-}
-
-#[cfg(test)]
-mod records_test {
-    use super::*;
-    use serde_json;
-
-    #[derive(Deserialize)]
-    struct Foo {
-        foo: String,
-    }
-
-    #[test]
-    fn it_parses_out_a_embedded_records_string() {
-        let json = r#"
-        {
-            "_embedded": {
-                "records": [
-                    { "foo": "bar" }
-                ]
-            }
-        }"#;
-        let records: Records<Foo> = serde_json::from_str(&json).unwrap();
-        assert_eq!(records.records().first().unwrap().foo, "bar");
     }
 }
